Forgot to download teh country codes.

+++ NCDCData/update_series_v2.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 yr=`date +%Y`
 mo=`date +%m`
 force=false
@@ -12,11 +12,17 @@ fi
 
 base=ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v2
 
+if [ ! -s v2.slp -o ! -s v2.slp.inv ]; then
+    echo "$0: error: v2.slp and v2.slp.inv should be procured elsewhere"
+    exit -1
+fi
+
 somethingnew=false
 
+wget -q -N $base/v2.prcp.inv
+wget -q -N $base/v2.country.codes
 for file in \
-    v2.prcp_adj.Z v2.prcp.Z \
-    v2.slp.Z
+    v2.prcp_adj.Z v2.prcp.Z # v2.slp.Z no longer exists at NCEI, last update was 2003, but we still offer an old version
 do
     cp $file $file.old
     wget -q -N $base/$file
